face_server/requirements.txt
click==8.0.3
dynaconf==3.1.9
face_recognition==1.3.0
Flask==2.1.2
Flask_Bcrypt==1.0.1
Flask_Cors==3.0.10
Flask_SQLAlchemy==2.5.1
PyJWT==2.4.0
pytest==7.1.2
setuptools==59.6.0
SQLAlchemy_serializer==1.4.1
